Separate Geocoder API from 'BragiPOI'
