+++ src/adapters/poi/bragi_poi.js
@@ -1,20 +1,4 @@
 import Poi from './poi';
-import ajax from '../../libs/ajax';
-import nconf from '@qwant/nconf-getter';
-import QueryContext from '../query_context';
-
-const serviceConfigs = nconf.get().services;
-const geocoderConfig = serviceConfigs.geocoder;
-const geocoderFocusPrecision = geocoderConfig.focusPrecision;
-
-if (!window.__bragiCache) {
-  window.__bragiCache = {};
-}
-
-function roundWithPrecision(value, precision) {
-  const rounded = Math.round(value * (1 / precision)) * precision;
-  return rounded.toFixed(3);
-}
 
 export default class BragiPoi extends Poi {
   constructor(feature, queryContext) {
@@ -87,69 +71,5 @@ export default class BragiPoi extends Poi {
     default:
       return this.name;
     }
-  }
-
-
-  static get(term, { lat, lon, zoom } = {}) {
-    let cacheKey = term;
-    if (lat !== undefined && lon !== undefined) {
-      lat = roundWithPrecision(lat, geocoderFocusPrecision);
-      lon = roundWithPrecision(lon, geocoderFocusPrecision);
-      cacheKey += `;${lat};${lon}`;
-    }
-    /* cache */
-    if (cacheKey in window.__bragiCache) {
-      const cachePromise = new Promise(resolve => {
-        resolve(window.__bragiCache[cacheKey]);
-      });
-      cachePromise.abort = () => {};
-      return cachePromise;
-    }
-
-    /* ajax */
-    let suggestsPromise;
-    const queryPromise = new Promise(async (resolve, reject) => {
-      const query = {
-        'q': term,
-        'limit': geocoderConfig.maxItems,
-      };
-      if (lat !== undefined && lon !== undefined) {
-        query.lat = lat;
-        query.lon = lon;
-      }
-      if (geocoderConfig.useLang) {
-        query.lang = window.getLang().code;
-      }
-      if (geocoderConfig.useNlu) {
-        query.nlu = 'true';
-      }
-      suggestsPromise = ajax.get(geocoderConfig.url, query);
-      suggestsPromise.then(suggests => {
-        let ranking = 0;
-        const bragiResponse = suggests.features.map(feature => {
-          ranking += 1;
-          const queryContext = new QueryContext(
-            term,
-            ranking,
-            query.lang,
-            { lat, lon, zoom }
-          );
-          return new BragiPoi(feature, queryContext);
-        });
-        window.__bragiCache[cacheKey] = bragiResponse;
-        resolve(bragiResponse);
-      }).catch(error => {
-        if (error === 0) { /* abort */
-          resolve(null);
-        } else {
-          reject(error);
-        }
-      });
-    });
-    queryPromise.abort = () => {
-      suggestsPromise.abort();
-    };
-
-    return queryPromise;
   }
 }
